+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 * #1 Optionals arguments not correctly interpreted
 * correct status command output with specified instance
+* fix: [System.Char] do not contains  « Trim » named method
 * Prepare v1.0.2