+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * #1 Optionals arguments not correctly interpreted
 * correct status command output with specified instance
 * fix: [System.Char] do not contains  « Trim » named method
+* add exec method  on help message
+* add sh method
 * Prepare v1.0.2